+++ Audio2Tag/Tag Manage/Main/Search/Album/TagSearchView+Extension.swift
@@ -1,0 +1,33 @@
+//
+//  TagSearchView+Extension.swift
+//  Audio2Tag
+//
+//  Created by Aoikazto on 2020/09/07.
+//  Copyright © 2020 Aoikazto. All rights reserved.
+//
+
+import SwiftUI
+
+extension UIApplication {
+    func endEditing(_ force: Bool) {
+        self.windows
+            .filter{$0.isKeyWindow}
+            .first?
+            .endEditing(force)
+    }
+}
+
+struct ResignKeyboardOnDragGesture: ViewModifier {
+    var gesture = DragGesture().onChanged{_ in
+        UIApplication.shared.endEditing(true)
+    }
+    func body(content: Content) -> some View {
+        content.gesture(gesture)
+    }
+}
+
+extension View {
+    func resignKeyboardOnDragGesture() -> some View {
+        return modifier(ResignKeyboardOnDragGesture())
+    }
+}
