+++ Audio2Tag/Tag Manage/Main/Search/Album/TagSearchView.swift
@@ -13,29 +13,6 @@ enum TagSearchKind {
     case musicBrainz
 }
 
-extension UIApplication {
-    func endEditing(_ force: Bool) {
-        self.windows
-            .filter{$0.isKeyWindow}
-            .first?
-            .endEditing(force)
-    }
-}
-
-struct ResignKeyboardOnDragGesture: ViewModifier {
-    var gesture = DragGesture().onChanged{_ in
-        UIApplication.shared.endEditing(true)
-    }
-    func body(content: Content) -> some View {
-        content.gesture(gesture)
-    }
-}
-
-extension View {
-    func resignKeyboardOnDragGesture() -> some View {
-        return modifier(ResignKeyboardOnDragGesture())
-    }
-}
 
 struct TagSearchView: View {
     @ObservedObject var viewModel = TagSearchViewModel()
